feature: fix error wapper variable

+++ src/main/java/com/euflausino/desafio_backend_it/dto/TransacaoRequestDTO.java
@@ -11,7 +11,7 @@ public record TransacaoRequestDTO(
 		
 		@NotNull
 		@Min(0)
-		BigDecimal valor,
+		Double valor,
 		
 		@NotNull
 		@PastOrPresent
